Remove popup display
