+++ frontend/app/(app)/edit-profile/photo-placeholder.tsx
