+++ frontend/design-system/components/shared/settings/setting.tsx
@@ -4,7 +4,11 @@ import { TextButton } from "@/design-system/components/shared/buttons/text-butto
 import { useUserStore } from "@/auth/store";
 import RedTextButton from "../buttons/red-text-button";
 
-const SettingContent = () => {
+interface SettingContentProps {
+  close: () => void;
+}
+
+const SettingContent = ({ close }: SettingContentProps) => {
   const { group } = useUserStore();
 
   // TODO: add routing for the setting menu
@@ -14,19 +18,28 @@ const SettingContent = () => {
         <Box gap="s">
           <TextButton
             textVariant="bodyLargeBold"
-            onPress={() => router.push("/(app)/(tabs)/profile/edit-profile")}
+            onPress={() => {
+              close();
+              router.push("/(app)/(tabs)/profile/edit-profile");
+            }}
             label="Edit Profile"
             variant="text"
           />
           <TextButton
             textVariant="bodyLargeBold"
-            onPress={() => router.push("/(app)/notification/config")}
+            onPress={() => {
+              close();
+              router.push("/(app)/notification/config");
+            }}
             label="Notifications"
             variant="text"
           />
           <TextButton
             textVariant="bodyLargeBold"
-            onPress={() => router.push("/(app)/group/member")}
+            onPress={() => {
+              close();
+              router.push("/(app)/group/member");
+            }}
             label="View Group"
             variant="text"
           />
